Update and clean code - Add test on persistCommand method

+++ src/main/java/ctoutweb/lalamiam/repository/StoreDayScheduleRepository.java
@@ -6,11 +6,10 @@ import ctoutweb.lalamiam.repository.entity.WeekDayEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface StoreDayScheduleRepository extends JpaRepository<StoreDayScheduleEntity, BigInteger> {
+public interface StoreDayScheduleRepository extends JpaRepository<StoreDayScheduleEntity, Long> {
   //List<ScheduleEntity> findAllByStore(StoreEntity store);
   List<StoreDayScheduleEntity> findAllByStoreAndWeekDay(StoreEntity store, WeekDayEntity weekDay);
 }
